+++ src/Services/CRM/Currency/Events/OnCrmCurrencyUpdate/OnCrmCurrencyUpdate.php
@@ -1,0 +1,26 @@
+<?php
+
+/**
+ * This file is part of the bitrix24-php-sdk package.
+ *
+ * © Vadim Soluyanov <vadimsallee@gmail.com>
+ *
+ * For the full copyright and license information, please view the MIT-LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Services\CRM\Currency\Events\OnCrmCurrencyUpdate;
+
+use Bitrix24\SDK\Application\Requests\Events\AbstractEventRequest;
+
+class OnCrmCurrencyUpdate extends AbstractEventRequest
+{
+    public const CODE = 'ONCRMCURRENCYUPDATE';
+
+    public function getPayload(): OnCrmCurrencyUpdatePayload
+    {
+        return new OnCrmCurrencyUpdatePayload($this->eventPayload['data']);
+    }
+}